+++ sdk/resourcemanager/Azure.ResourceManager/tests/Samples/Readme.cs
@@ -1,13 +1,13 @@
 ﻿#region Snippet:Readme_AuthClient
 using Azure.Identity;
-using Azure.ResourceManager.Core;
+using Azure.ResourceManager;
+using Azure.ResourceManager.Resources;
 using System;
 using System.Threading.Tasks;
-using Azure.ResourceManager.Resources;
 #if !SNIPPET
 using NUnit.Framework;
 
-namespace Azure.ResourceManager.Core.Tests.Samples
+namespace Azure.ResourceManager.Tests.Samples
 {
     class Readme
     {
@@ -19,7 +19,7 @@ namespace Azure.ResourceManager.Core.Tests.Samples
 
 // Code omitted for brevity
 
-var armClient = new ArmClient(new DefaultAzureCredential());
+ArmClient armClient = new ArmClient(new DefaultAzureCredential());
 #endregion Snippet:Readme_AuthClient
         }
 
@@ -30,7 +30,7 @@ var armClient = new ArmClient(new DefaultAzureCredential());
             #region Snippet:Readme_CastToSpecificType
             string resourceId = "/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/workshop2021-rg/providers/Microsoft.Network/virtualNetworks/myVnet/subnets/mySubnet";
             // We know the subnet is a resource group level identifier since it has a resource group name in its string
-            ResourceGroupResourceIdentifier id = resourceId;
+            ResourceIdentifier id = new ResourceIdentifier(resourceId);
             Console.WriteLine($"Subscription: {id.SubscriptionId}");
             Console.WriteLine($"ResourceGroup: {id.ResourceGroupName}");
             Console.WriteLine($"Vnet: {id.Parent.Name}");
@@ -45,7 +45,7 @@ var armClient = new ArmClient(new DefaultAzureCredential());
             #region Snippet:Readme_CastToBaseResourceIdentifier
             string resourceId = "/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/workshop2021-rg/providers/Microsoft.Network/virtualNetworks/myVnet/subnets/mySubnet";
             // Assume we don't know what type of resource id we have we can cast to the base type
-            ResourceIdentifier id = resourceId;
+            ResourceIdentifier id = new ResourceIdentifier(resourceId);
             string property;
             if (id.TryGetSubscriptionId(out property))
                 Console.WriteLine($"Subscription: {property}");
@@ -61,18 +61,18 @@ var armClient = new ArmClient(new DefaultAzureCredential());
         public async Task CheckIfResourceGroupExists()
         {
             #region Snippet:Readme_CheckIfExistssRG
-            var armClient = new ArmClient(new DefaultAzureCredential());
+            ArmClient armClient = new ArmClient(new DefaultAzureCredential());
             Subscription subscription = armClient.DefaultSubscription;
             string rgName = "myRgName";
 
-            var exists = await subscription.GetResourceGroups().CheckIfExistsAsync(rgName);
+            bool exists = await subscription.GetResourceGroups().CheckIfExistsAsync(rgName);
 
             if (exists)
             {
                 Console.WriteLine($"Resource Group {rgName} exists.");
 
                 // We can get the resource group now that we are sure it exists.
-                var myRG = await subscription.GetResourceGroups().GetAsync(rgName);
+                ResourceGroup myRG = await subscription.GetResourceGroups().GetAsync(rgName);
             }
             else
             {
@@ -86,11 +86,11 @@ var armClient = new ArmClient(new DefaultAzureCredential());
         public async Task TryGetResourceGroup()
         {
             #region Snippet:Readme_TryGetRG
-            var armClient = new ArmClient(new DefaultAzureCredential());
+            ArmClient armClient = new ArmClient(new DefaultAzureCredential());
             Subscription subscription = armClient.DefaultSubscription;
             string rgName = "myRgName";
 
-            var myRG = await subscription.GetResourceGroups().TryGetAsync(rgName);
+            ResourceGroup myRG = await subscription.GetResourceGroups().GetIfExistsAsync(rgName);
 
             if (myRG == null)
             {