+++ sdk/resourcemanager/Azure.ResourceManager/tests/Unit/WritableSubResourceTests.cs
@@ -3,7 +3,7 @@ using Azure.ResourceManager.Resources.Models;
 using Azure.ResourceManager.TestFramework;
 using NUnit.Framework;
 
-namespace Azure.ResourceManager.Core.Tests
+namespace Azure.ResourceManager.Tests
 {
     [Parallelizable]
     public class WritableSubResourceTests